refactor(modals): simplify/clean open/close functions
