+++ assets/javascript/index.js
@@ -1,15 +1,22 @@
-import closeModal from './modules/closeModals.js';
-import openContact from './modules/openContact.js';
-import openLightbox from './modules/openLightbox.js';
+import closeModal from './modules/modals/closeModals.js';
+import openModal from './modules/modals/openModal.js';
 
+import focusOutInputCheck from './modules/modals/contact/focusOutInputCheck.js';
+
+// modals
 const contactModalButton = document.getElementById('contact-button');
 const contactModalBackground = document.getElementById('contact-form__background');
 const lightboxModalButtons = document.querySelectorAll('.portfolio__element-media');
 const lightboxModalBackground = document.getElementById('lightbox__background');
 const closeModalsButtons = document.querySelectorAll('.close-button');
 
-contactModalButton.addEventListener('click', () => { openContact(contactModalBackground); });
+contactModalButton.addEventListener('click', () => { openModal(contactModalBackground); });
 
-lightboxModalButtons.forEach((btn) => btn.addEventListener('click', () => { openLightbox(lightboxModalBackground); }));
+lightboxModalButtons.forEach((btn) => btn.addEventListener('click', () => { openModal(lightboxModalBackground); }));
 
 closeModalsButtons.forEach((btn) => btn.addEventListener('click', () => { closeModal(contactModalBackground, lightboxModalBackground); }));
+
+// contact modal validation
+const contactFormInputs = document.querySelectorAll('input.contact-form__text-input, textarea.contact-form__text-input');
+
+contactFormInputs.forEach((input) => input.addEventListener('focusout', () => { focusOutInputCheck(input); }));
