+++ assets/javascript/modules/modals/closeModals.js
