+++ assets/javascript/modules/modals/openModal.js
@@ -1,0 +1,5 @@
+function openModal(modalBackground) {
+    modalBackground.style.display = 'unset';
+}
+
+export default openModal;
