--- assets/javascript/modules/openContact.js
@@ -1,5 +1,0 @@
-function openContact(contactModalBackground) {
-    contactModalBackground.style.display = 'unset';
-}
-
-export default openContact;
