--- assets/javascript/modules/openLightbox.js
@@ -1,5 +1,0 @@
-function openLightbox(lightboxModalBackground) {
-    lightboxModalBackground.style.display = 'unset';
-}
-
-export default openLightbox;
